Reorganize dashboard layout and task list
